backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
gunicorn
dnspython
pyjwt
werkzeug
google-generativeai
PyPDF2